# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sphere Platform API",
    "version": "4.0.0"
  },
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login: получить access token + refresh cookie",
        "description": "Аутентификация пользователя.\n- Если MFA отключён: возвращает access_token + устанавливает HTTPOnly refresh_token cookie.\n- Если MFA включён: возвращает mfa_required=True + state_token (не выдаёт JWT до подтверждения).\nRate limit: 5 попыток с IP за 60 секунд → 429.",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/LoginRequest"
                  }
                ],
                "title": "Body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/TokenResponse"
                    },
                    {
                      "$ref": "#/components/schemas/MFARequiredResponse"
                    }
                  ],
                  "title": "Response Login Api V1 Auth Login Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login/mfa": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Второй шаг MFA login: подтвердить TOTP-код",
        "description": "Второй шаг MFA: принимает state_token (из /login) и 6-значный TOTP-код.\nПри успехе выдаёт access_token + refresh_token cookie.",
        "operationId": "login_mfa_api_v1_auth_login_mfa_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/MFALoginRequest"
                  }
                ],
                "title": "Body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Обновить access token по refresh cookie",
        "description": "Обновить access token по HTTPOnly refresh_token cookie.\nРотация: старый refresh token инвалидируется, выпускается новый.",
        "operationId": "refresh_api_v1_auth_refresh_post",
        "parameters": [
          {
            "name": "refresh_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Refresh Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Logout: инвалидировать токены",
        "description": "Logout: помещает access token в Redis blacklist + отзывает refresh token в БД.\nFIX-1.4: использует decode_expired_access_token — работает даже с просроченным токеном.\nВсегда удаляет refresh_token cookie независимо от результата.",
        "operationId": "logout_api_v1_auth_logout_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "refresh_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Refresh Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Информация о текущем пользователе",
        "description": "Получить профиль текущего аутентифицированного пользователя.",
        "operationId": "me_api_v1_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/auth/mfa/setup": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "SPLIT-2: Шаг 1 — Сгенерировать TOTP QR-код",
        "description": "Шаг 1 MFA setup: сгенерировать TOTP секрет и QR-код.\nСекрет временно сохраняется в Redis (5 минут) — до подтверждения через /mfa/verify-setup.",
        "operationId": "mfa_setup_api_v1_auth_mfa_setup_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MFASetupResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/auth/mfa/verify-setup": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "SPLIT-2: Шаг 2 — Подтвердить TOTP-код и включить MFA",
        "description": "Шаг 2 MFA setup: проверить отсканированный QR-код и активировать MFA.\nБез этого подтверждения MFA не включается — QR-код может быть неверно отсканирован.",
        "operationId": "mfa_verify_setup_api_v1_auth_mfa_verify_setup_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/MFAVerifySetupRequest"
                  }
                ],
                "title": "Body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/auth/mfa": {
      "delete": {
        "tags": [
          "auth"
        ],
        "summary": "SPLIT-2: Отключить MFA",
        "description": "Отключить TOTP MFA для текущего пользователя.",
        "operationId": "mfa_disable_api_v1_auth_mfa_delete",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/auth/api-keys": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "SPLIT-4: Список API ключей",
        "description": "Список активных API ключей для org текущего пользователя.",
        "operationId": "list_api_keys_api_v1_auth_api_keys_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/APIKeyResponse"
                  },
                  "type": "array",
                  "title": "Response List Api Keys Api V1 Auth Api Keys Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "SPLIT-4: Создать API ключ",
        "description": "Создать новый API ключ для машинных интеграций.\nraw_key показывается ОДИН РАЗ в ответе — после этого получить его невозможно.",
        "operationId": "create_api_key_api_v1_auth_api_keys_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/CreateAPIKeyRequest"
                  }
                ],
                "title": "Body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIKeyCreatedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/auth/api-keys/{key_id}": {
      "delete": {
        "tags": [
          "auth"
        ],
        "summary": "SPLIT-4: Отозвать API ключ",
        "description": "Отозвать (деактивировать) API ключ.",
        "operationId": "revoke_api_key_api_v1_auth_api_keys__key_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "key_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Key Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "APIKeyCreatedResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "key_prefix": {
            "type": "string",
            "title": "Key Prefix"
          },
          "raw_key": {
            "type": "string",
            "title": "Raw Key"
          },
          "permissions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Permissions"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "key_prefix",
          "raw_key",
          "permissions",
          "created_at"
        ],
        "title": "APIKeyCreatedResponse"
      },
      "APIKeyResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "key_prefix": {
            "type": "string",
            "title": "Key Prefix"
          },
          "permissions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Permissions"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "last_used_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Used At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "key_prefix",
          "permissions",
          "is_active",
          "created_at"
        ],
        "title": "APIKeyResponse",
        "description": "Список API ключей (raw_key НЕ включается)."
      },
...
      "CreateAPIKeyRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name"
          },
          "permissions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Permissions"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "key_type": {
            "type": "string",
            "pattern": "^(user|agent)$",
            "title": "Key Type",
            "default": "user"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "CreateAPIKeyRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoginRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 8,
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest"
      },
      "MFALoginRequest": {
        "properties": {
          "state_token": {
            "type": "string",
            "title": "State Token"
          },
          "code": {
            "type": "string",
            "maxLength": 8,
            "minLength": 6,
            "title": "Code"
          }
        },
        "type": "object",
        "required": [
          "state_token",
          "code"
        ],
        "title": "MFALoginRequest"
      },
      "MFARequiredResponse": {
        "properties": {
          "mfa_required": {
            "type": "boolean",
            "title": "Mfa Required",
            "default": true
          },
          "state_token": {
            "type": "string",
            "title": "State Token"
          }
        },
        "type": "object",
        "required": [
          "state_token"
        ],
        "title": "MFARequiredResponse"
      },
      "MFASetupResponse": {
        "properties": {
          "qr_code": {
            "type": "string",
            "title": "Qr Code"
          },
          "secret": {
            "type": "string",
            "title": "Secret"
          }
        },
        "type": "object",
        "required": [
          "qr_code",
          "secret"
        ],
        "title": "MFASetupResponse"
      },
      "MFAVerifySetupRequest": {
        "properties": {
          "code": {
            "type": "string",
            "maxLength": 8,
            "minLength": 6,
            "title": "Code"
          }
        },
        "type": "object",
        "required": [
          "code"
        ],
        "title": "MFAVerifySetupRequest"
      },
...
      "TokenResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "expires_in": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires In"
          }
        },
        "type": "object",
        "required": [
          "access_token"
        ],
        "title": "TokenResponse"
      },
...
      "UserResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "org_id": {
            "type": "string",
            "format": "uuid",
            "title": "Org Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "mfa_enabled": {
            "type": "boolean",
            "title": "Mfa Enabled"
          },
          "last_login_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Login At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "org_id",
          "email",
          "role",
          "is_active",
          "mfa_enabled",
          "created_at"
        ],
        "title": "UserResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}